+++ samples/webview/mac/helper-Info.plist
