+++ samples/webview/mac/tools/change_mach_o_flags.py
