+++ samples/webview/mac/tools/make_more_helpers.sh
