+++ samples/webview/mac/webview_chromium-Info.plist
