--- samples/webview_chromium/refresh.xpm
@@ -1,234 +1,0 @@
-/* XPM */
-static const char * refresh_xpm[] = {
-"24 24 207 2",
-"  	c None",
-". 	c #3162A2",
-"+ 	c #507BB3",
-"@ 	c #6A90C0",
-"# 	c #7296C4",
-"$ 	c #7397C4",
-"% 	c #7094C2",
-"& 	c #648BBD",
-"* 	c #507CB3",
-"= 	c #3061A2",
-"- 	c #8EACD1",
-"; 	c #B3CBE4",
-"> 	c #B7CEE6",
-", 	c #B8CEE6",
-"' 	c #B5CDE5",
-") 	c #B4CBE5",
-"! 	c #B0C9E3",
-"~ 	c #AEC6E1",
-"{ 	c #9FBBDB",
-"] 	c #648DBE",
-"^ 	c #2F60A2",
-"/ 	c #3262A4",
-"( 	c #3364A3",
-"_ 	c #3C6BA8",
-": 	c #A7C0DD",
-"< 	c #C1D6EA",
-"[ 	c #BAD0E7",
-"} 	c #B6CDE5",
-"| 	c #B6CCE6",
-"1 	c #B5CDE6",
-"2 	c #B3CBE5",
-"3 	c #AFC8E3",
-"4 	c #AAC5E2",
-"5 	c #ABC6E2",
-"6 	c #A0BCDB",
-"7 	c #4975AF",
-"8 	c #3666A5",
-"9 	c #89ACD3",
-"0 	c #2C5EA0",
-"a 	c #82A3CB",
-"b 	c #B5CCE4",
-"c 	c #7095C3",
-"d 	c #658CBC",
-"e 	c #678EBE",
-"f 	c #7196C2",
-"g 	c #84A4CC",
-"h 	c #A9C3E0",
-"i 	c #B8CFE7",
-"j 	c #ACC6E3",
-"k 	c #A0BEDE",
-"l 	c #9CBCDD",
-"m 	c #9EBDDE",
-"n 	c #6C93C1",
-"o 	c #7B9FCA",
-"p 	c #ABC7E3",
-"q 	c #2B5DA0",
-"r 	c #4976AE",
-"s 	c #8EADD0",
-"t 	c #4975AE",
-"u 	c #3667A6",
-"v 	c #3666A6",
-"w 	c #4170AB",
-"x 	c #5B83B6",
-"y 	c #9BB9DA",
-"z 	c #97B9DC",
-"A 	c #7EA7D3",
-"B 	c #7CA5D2",
-"C 	c #88AED6",
-"D 	c #8AAFD6",
-"E 	c #A4C1E0",
-"F 	c #406DAB",
-"G 	c #4F7BB2",
-"H 	c #4A77B3",
-"I 	c #4C79B4",
-"J 	c #2E5EA2",
-"K 	c #6F95C3",
-"L 	c #95B6D9",
-"M 	c #7FA8D3",
-"N 	c #76A1D0",
-"O 	c #709DCE",
-"P 	c #2B5DA1",
-"Q 	c #3667A5",
-"R 	c #4D7AB4",
-"S 	c #648DC3",
-"T 	c #4A78B3",
-"U 	c #3262A2",
-"V 	c #2C5FA0",
-"W 	c #6B91C0",
-"X 	c #97B8DB",
-"Y 	c #74A0CF",
-"Z 	c #709ECE",
-"` 	c #ACC7E3",
-" .	c #2D5EA0",
-"..	c #3767A6",
-"+.	c #5985BD",
-"@.	c #78A3D0",
-"#.	c #6C9BCC",
-"$.	c #AEC9E5",
-"%.	c #3263A3",
-"&.	c #3768A7",
-"*.	c #9BB8DA",
-"=.	c #9CB9DA",
-"-.	c #9CBADA",
-";.	c #9FBBDC",
-">.	c #A8C1DF",
-",.	c #2D5FA0",
-"'.	c #22569B",
-").	c #23579C",
-"!.	c #21569A",
-"~.	c #21559B",
-"{.	c #275A9E",
-"].	c #3364A4",
-"^.	c #24589C",
-"/.	c #24579C",
-"(.	c #416FAB",
-"_.	c #C0D3E8",
-":.	c #B8CCE4",
-"<.	c #B5CBE3",
-"[.	c #AFC7E1",
-"}.	c #B0C8E2",
-"|.	c #88A9CF",
-"1.	c #4471AC",
-"2.	c #B9D1E8",
-"3.	c #94B6DA",
-"4.	c #95B7DB",
-"5.	c #A3C1DF",
-"6.	c #A4BEDC",
-"7.	c #2D60A0",
-"8.	c #BBD1E9",
-"9.	c #BCD1E8",
-"0.	c #416EA9",
-"a.	c #3868A7",
-"b.	c #BAD1E9",
-"c.	c #96B7DB",
-"d.	c #98B9DC",
-"e.	c #A2C0DF",
-"f.	c #AAC4DF",
-"g.	c #5F88BA",
-"h.	c #80A8D7",
-"i.	c #CDDCEC",
-"j.	c #7A9FC8",
-"k.	c #4273B1",
-"l.	c #3465A4",
-"m.	c #4371AC",
-"n.	c #B9D1E9",
-"o.	c #B4CCE5",
-"p.	c #9DBCDD",
-"q.	c #8EB2D8",
-"r.	c #91B3D9",
-"s.	c #719AC8",
-"t.	c #4171AC",
-"u.	c #3A6AA8",
-"v.	c #3A75BD",
-"w.	c #9CBBDB",
-"x.	c #608BC1",
-"y.	c #3A6EAD",
-"z.	c #4C80BB",
-"A.	c #3365A4",
-"B.	c #C6DAED",
-"C.	c #6F94C1",
-"D.	c #8CABD1",
-"E.	c #9BBBDD",
-"F.	c #6C9ACD",
-"G.	c #6596CA",
-"H.	c #709DCF",
-"I.	c #7AA4D2",
-"J.	c #6A96C8",
-"K.	c #5885BB",
-"L.	c #4F7EB7",
-"M.	c #497AB6",
-"N.	c #487AB6",
-"O.	c #4275B5",
-"P.	c #4C7FBB",
-"Q.	c #578BC6",
-"R.	c #396CAA",
-"S.	c #789CC6",
-"T.	c #5681B6",
-"U.	c #8AAFD7",
-"V.	c #77A2D1",
-"W.	c #6496CA",
-"X.	c #6294C9",
-"Y.	c #6395CA",
-"Z.	c #5F92C9",
-"`.	c #5C8FC7",
-" +	c #598DC6",
-".+	c #588DC7",
-"++	c #4175B2",
-"@+	c #31619F",
-"#+	c #6089BB",
-"$+	c #77A0CE",
-"%+	c #6E9CCD",
-"&+	c #6696CA",
-"*+	c #6193C9",
-"=+	c #5F91C8",
-"-+	c #5C90C7",
-";+	c #5A8EC6",
-">+	c #558AC3",
-",+	c #4072AF",
-"'+	c #27528A",
-")+	c #2A5285",
-"!+	c #3C6599",
-"~+	c #4B77B0",
-"{+	c #4E7DB6",
-"]+	c #4C7CB7",
-"^+	c #4877B2",
-"/+	c #3B69A2",
-"(+	c #3361A0",
-"                                                ",
-"            . + @ # $ % & *                     ",
-"          = - ; > , ' ) ! ~ { ] ^       / (     ",
-"        _ : < [ } | 1 2 3 4 5 > 6 7   8 9 0     ",
-"        a b a c d e f g h i j k l m n o p q     ",
-"      r s t u v         w x y z A B C D E q     ",
-"      F G H I               J K L M N O 4 P     ",
-"      Q R S T U               V W X Y Z `  .    ",
-"      ..+.                    t L @.Z #.$.0     ",
-"      %.                    &.{ *.=.-.;.>.,.    ",
-"        '.).!.~.~.{.].      %.^./.^./.).'.      ",
-"      (._.:.<.[.}.|.V                     ].    ",
-"      1.2.3.4.5.6.7.                      %.    ",
-"      1.8.X l 9.0.                      a.(     ",
-"      1.b.c.d.e.f.g.              h.i.j.k.l.    ",
-"      m.n.o.p.3.q.r.s.t.u.      v.w.x.y.z.A.    ",
-"      m.B.C.D.E.F.G.H.I.J.K.L.M.N.O.P.Q.R.      ",
-"      (.S.    T.U.V.W.X.Y.Y.X.Z.`. +.+++        ",
-"      ].        @+#+$+%+&+*+=+-+;+>+,+          ",
-"                  '+)+!+~+{+]+^+/+(+            ",
-"                                                ",
-"                                                ",
-"                                                ",
-"                                                "};
