--- samples/webview_chromium/stop.xpm
@@ -1,331 +1,0 @@
-/* XPM */
-static const char * stop_xpm[] = {
-"24 24 304 2",
-"  	c None",
-". 	c #920F0F",
-"+ 	c #931212",
-"@ 	c #931111",
-"# 	c #8F0909",
-"$ 	c #981414",
-"% 	c #DC6666",
-"& 	c #F07575",
-"* 	c #EF7272",
-"= 	c #EE7171",
-"- 	c #EE6F6F",
-"; 	c #EE6D6D",
-"> 	c #EE6C6C",
-", 	c #E46969",
-"' 	c #980F0F",
-") 	c #8E0606",
-"! 	c #EA7B7B",
-"~ 	c #E25858",
-"{ 	c #D84040",
-"] 	c #D83F3F",
-"^ 	c #D83E3E",
-"/ 	c #D83D3D",
-"( 	c #D73C3C",
-"_ 	c #D73A3A",
-": 	c #DD4A4A",
-"< 	c #EC7272",
-"[ 	c #950909",
-"} 	c #9A1616",
-"| 	c #EB7C7C",
-"1 	c #E15959",
-"2 	c #D74040",
-"3 	c #D94444",
-"4 	c #D94343",
-"5 	c #D94242",
-"6 	c #D94141",
-"7 	c #D93F3F",
-"8 	c #D73D3D",
-"9 	c #D73939",
-"0 	c #DD4747",
-"a 	c #E96D6D",
-"b 	c #A21818",
-"c 	c #9D1717",
-"d 	c #DC6868",
-"e 	c #E25959",
-"f 	c #D94545",
-"g 	c #D84545",
-"h 	c #DA4545",
-"i 	c #D84242",
-"j 	c #D84141",
-"k 	c #D63D3D",
-"l 	c #D73B3B",
-"m 	c #D63838",
-"n 	c #DC4747",
-"o 	c #D95555",
-"p 	c #9D1414",
-"q 	c #940909",
-"r 	c #E15757",
-"s 	c #DC4848",
-"t 	c #F64949",
-"u 	c #D94747",
-"v 	c #D94646",
-"w 	c #D73F3F",
-"x 	c #EA4141",
-"y 	c #E53D3D",
-"z 	c #D53A3A",
-"A 	c #D43838",
-"B 	c #DB4747",
-"C 	c #E66161",
-"D 	c #920606",
-"E 	c #9D1818",
-"F 	c #EC7C7C",
-"G 	c #DF5353",
-"H 	c #DC4545",
-"I 	c #F15A5A",
-"J 	c #EEB2B2",
-"K 	c #F55757",
-"L 	c #DB4545",
-"M 	c #EA4545",
-"N 	c #F38383",
-"O 	c #F27979",
-"P 	c #DC3D3D",
-"Q 	c #D33838",
-"R 	c #D13535",
-"S 	c #D84343",
-"T 	c #E35C5C",
-"U 	c #980E0E",
-"V 	c #BD4949",
-"W 	c #E25656",
-"X 	c #D74343",
-"Y 	c #F44F4F",
-"Z 	c #EBBFBF",
-"` 	c #E4EBEB",
-" .	c #EBC7C7",
-"..	c #F15D5D",
-"+.	c #DB4141",
-"@.	c #EA4747",
-"#.	c #F09C9C",
-"$.	c #E4EFEF",
-"%.	c #E5DDDD",
-"&.	c #FC5858",
-"*.	c #D53939",
-"=.	c #D13737",
-"-.	c #CE3333",
-";.	c #B93434",
-">.	c #DC4949",
-",.	c #E34545",
-"'.	c #F48F8F",
-").	c #E7E6E6",
-"!.	c #E8E9E9",
-"~.	c #EDCBCB",
-"{.	c #F45A5A",
-"].	c #EA4C4C",
-"^.	c #F2A3A3",
-"/.	c #E9E9E9",
-"(.	c #E7E9E9",
-"_.	c #F0B0B0",
-":.	c #E74848",
-"<.	c #CF3636",
-"[.	c #CE3535",
-"}.	c #D23A3A",
-"|.	c #BC4949",
-"1.	c #DD4848",
-"2.	c #E74444",
-"3.	c #F29E9E",
-"4.	c #EAE9E9",
-"5.	c #EBF0F0",
-"6.	c #F0CECE",
-"7.	c #F79191",
-"8.	c #EAF4F4",
-"9.	c #EBECEC",
-"0.	c #EFC4C4",
-"a.	c #EA4B4B",
-"b.	c #D13939",
-"c.	c #CF3737",
-"d.	c #CD3434",
-"e.	c #D13A3A",
-"f.	c #B83333",
-"g.	c #D73E3E",
-"h.	c #E74646",
-"i.	c #F59E9E",
-"j.	c #EFE8E8",
-"k.	c #EEF1F1",
-"l.	c #EEF5F5",
-"m.	c #F3BABA",
-"n.	c #EA5050",
-"o.	c #D23B3B",
-"p.	c #D03A3A",
-"q.	c #CF3939",
-"r.	c #CC3535",
-"s.	c #CA3131",
-"t.	c #B83232",
-"u.	c #BD4848",
-"v.	c #D83A3A",
-"w.	c #F05D5D",
-"x.	c #F5BEBE",
-"y.	c #F1F2F2",
-"z.	c #F2E6E6",
-"A.	c #FE7A7A",
-"B.	c #CF3434",
-"C.	c #C92121",
-"D.	c #C61212",
-"E.	c #C10707",
-"F.	c #C00505",
-"G.	c #BE0404",
-"H.	c #C30C0C",
-"I.	c #B72222",
-"J.	c #BC4848",
-"K.	c #DC4343",
-"L.	c #D73838",
-"M.	c #E95A5A",
-"N.	c #F8ADAD",
-"O.	c #F2F9F9",
-"P.	c #F3F3F3",
-"Q.	c #F4F2F2",
-"R.	c #F3F7F7",
-"S.	c #F2D5D5",
-"T.	c #F05050",
-"U.	c #C30000",
-"V.	c #BF0000",
-"W.	c #BE0000",
-"X.	c #BD0000",
-"Y.	c #BB0000",
-"Z.	c #C00707",
-"`.	c #B71F1F",
-" +	c #BC4646",
-".+	c #DC4141",
-"++	c #D63636",
-"@+	c #D63737",
-"#+	c #EA5C5C",
-"$+	c #F6B9B9",
-"%+	c #F4FAFA",
-"&+	c #F9C5C5",
-"*+	c #FCACAC",
-"=+	c #F4F0F0",
-"-+	c #F1F6F6",
-";+	c #EB5F5F",
-">+	c #C00101",
-",+	c #B90000",
-"'+	c #BF0707",
-")+	c #B61D1D",
-"!+	c #BC4545",
-"~+	c #DA3E3E",
-"{+	c #D73535",
-"]+	c #D63434",
-"^+	c #EA6363",
-"/+	c #F5BFBF",
-"(+	c #F1F1F1",
-"_+	c #F5F6F6",
-":+	c #FAC9C9",
-"<+	c #DF3C3C",
-"[+	c #D21B1B",
-"}+	c #FBADAD",
-"|+	c #F5F4F4",
-"1+	c #F2F2F2",
-"2+	c #F1D7D7",
-"3+	c #F07373",
-"4+	c #B80000",
-"5+	c #BD0707",
-"6+	c #B61C1C",
-"7+	c #B53B3B",
-"8+	c #E75D5D",
-"9+	c #D63333",
-"0+	c #D53333",
-"a+	c #D42F2F",
-"b+	c #F58686",
-"c+	c #F3D4D4",
-"d+	c #F2F3F3",
-"e+	c #F6D3D3",
-"f+	c #E44343",
-"g+	c #C40000",
-"h+	c #C20000",
-"i+	c #D82C2C",
-"j+	c #F7B7B7",
-"k+	c #F0E9E9",
-"l+	c #FA8F8F",
-"m+	c #BE0A0A",
-"n+	c #B70000",
-"o+	c #B60000",
-"p+	c #C20C0C",
-"q+	c #B41919",
-"r+	c #8B0101",
-"s+	c #C44848",
-"t+	c #EC5F5F",
-"u+	c #D52A2A",
-"v+	c #D22424",
-"w+	c #CE1C1C",
-"x+	c #E96969",
-"y+	c #FBC4C4",
-"z+	c #E44949",
-"A+	c #C40101",
-"B+	c #C10000",
-"C+	c #D53131",
-"D+	c #FBBEBE",
-"E+	c #F09292",
-"F+	c #C00E0E",
-"G+	c #B50000",
-"H+	c #C71212",
-"I+	c #980707",
-"J+	c #890101",
-"K+	c #B52828",
-"L+	c #E63D3D",
-"M+	c #C90202",
-"N+	c #C70000",
-"O+	c #C80505",
-"P+	c #D32424",
-"Q+	c #C61B1B",
-"R+	c #C11111",
-"S+	c #C81313",
-"T+	c #8C0101",
-"U+	c #880000",
-"V+	c #BE2E2E",
-"W+	c #DC2D2D",
-"X+	c #C50000",
-"Y+	c #BC0000",
-"Z+	c #C30D0D",
-"`+	c #CE2424",
-" @	c #850000",
-".@	c #BE2B2B",
-"+@	c #E33434",
-"@@	c #C00000",
-"#@	c #CC1616",
-"$@	c #CB2222",
-"%@	c #910404",
-"&@	c #790000",
-"*@	c #B21D1D",
-"=@	c #E03131",
-"-@	c #B40000",
-";@	c #CD1818",
-">@	c #C61F1F",
-",@	c #830000",
-"'@	c #730000",
-")@	c #B92525",
-"!@	c #E14444",
-"~@	c #E04141",
-"{@	c #DF3F3F",
-"]@	c #DE3C3C",
-"^@	c #DE3A3A",
-"/@	c #DE3737",
-"(@	c #DD3535",
-"_@	c #DD3232",
-":@	c #C42121",
-"<@	c #7E0000",
-"[@	c #720000",
-"}@	c #770000",
-"              . + + + + + + @ @ #               ",
-"            $ % & * * * = - ; > , '             ",
-"          ) ! ~ { { { ] ^ / ( _ : < [           ",
-"        } | 1 2 3 3 4 5 6 7 ^ 8 9 0 a b         ",
-"      c d e 6 f g f h 4 i j 7 k l m n o p       ",
-"    q | r 4 f s t u v h i 6 w x y z A B C D     ",
-"  E F G w 4 H I J K L f i j M N O P Q R S T U   ",
-"  V W / 6 X Y Z `  ...+.j @.#.$.%.&.*.=.-.f ;.  ",
-"  V >.7 { i ,.'.).!.~.{.].^./.(._.:.=.<.[.}.;.  ",
-"  |.1.^ ] 7 { 2.3.4.5.6.7.8.9.0.a.b.c.[.d.e.f.  ",
-"  |.0 ( 8 ^ ] g.h.i.j.k.l.k.m.n.o.p.q.r.s.<.t.  ",
-"  u.H v.l ( / / *.w.x.y.y.z.A.B.C.D.E.F.G.H.I.  ",
-"  J.K.L.9 _ _ _ M.N.O.P.Q.R.S.T.U.V.W.X.Y.Z.`.  ",
-"   +.+++@+L.@+#+$+O.%+&+*+=+-+S.;+>+X.Y.,+'+)+  ",
-"  !+~+{+{+]+^+/+(+_+:+<+[+}+|+1+2+3+Y.,+4+5+6+  ",
-"  7+8+9+0+a+b+c+d+e+f+g+h+i+j+P.k+l+m+n+o+p+q+  ",
-"  r+s+t+u+v+w+x+y+z+A+h+B+X.C+D+E+F+n+G+H+v+I+  ",
-"    J+K+L+M+N+O+P+A+h+B+V.X.Y.Q+R+4+o+S+v+T+    ",
-"      U+V+W+N+X+U.h+B+V.X.Y+Y.,+4+o+Z+`+T+      ",
-"         @.@+@A+h+@@W.X.Y+Y.,+4+G+#@$@%@        ",
-"          &@*@=@V.Y+Y.,+4+o+G+-@;@>@,@          ",
-"            '@)@!@~@{@]@^@/@(@_@:@<@            ",
-"              [@&@&@&@&@&@&@&@&@}@              ",
-"                                                "};
