--- samples/webview_chromium/wxlogo.xpm
@@ -1,84 +1,0 @@
-/* XPM */
-static const char * wxlogo_xpm[] = {
-"24 24 57 1",
-" 	c None",
-".	c #000006",
-"+	c #00003F",
-"@	c #00003C",
-"#	c #000034",
-"$	c #000000",
-"%	c #00000E",
-"&	c #000085",
-"*	c #000079",
-"=	c #000081",
-"-	c #000070",
-";	c #00000D",
-">	c #8686CB",
-",	c #FFFFFF",
-"'	c #000078",
-")	c #000080",
-"!	c #00006F",
-"~	c #7F7FC7",
-"{	c #00007F",
-"]	c #000083",
-"^	c #000088",
-"/	c #000076",
-"(	c #000047",
-"_	c #000040",
-":	c #EDED00",
-"<	c #FFFF00",
-"[	c #202000",
-"}	c #0E0000",
-"|	c #7F0000",
-"1	c #770000",
-"2	c #7E0000",
-"3	c #E0E000",
-"4	c #FFFF88",
-"5	c #FFFF3F",
-"6	c #1E1E00",
-"7	c #1F0000",
-"8	c #FF0000",
-"9	c #FFFF80",
-"0	c #1E0000",
-"a	c #FF8686",
-"b	c #00000F",
-"c	c #000082",
-"d	c #FFFF7F",
-"e	c #FF7F7F",
-"f	c #000007",
-"g	c #000044",
-"h	c #00003E",
-"i	c #00001F",
-"j	c #DE0000",
-"k	c #E00000",
-"l	c #C30000",
-"m	c #DEDE00",
-"n	c #DD0000",
-"o	c #DDDD00",
-"p	c #EAEA00",
-"q	c #780000",
-"r	c #680000",
-"                        ",
-"                        ",
-"                        ",
-"                        ",
-"      .+@@@@@@@#$       ",
-"      %&*======-$       ",
-"      ;>,')))))!$       ",
-"      ;~,')){]^/        ",
-"      ;~,'))=(%;$$$$$$$$",
-"      ;~,'))=_:<<<<<<<<[",
-"}|1112;~,'))=_345<<<<<<6",
-"788888;>,'))=_3,9<<<<<<6",
-"0a,888b&*===c_3,d<<<<<<6",
-"0e,888fg__h@@i3,d<<<<<<6",
-"0e,888jkkl$  $m,d<<<<<<6",
-"0e,888888n$  $o,9<<<<<<6",
-"0e,888888n$  $o45<<<<<<6",
-"0a,888888n$  $p<<<<<<<<[",
-"788888888k$   $$$$$$$$$$",
-"}|qqqqqqqr$             ",
-"                        ",
-"                        ",
-"                        ",
-"                        "};
